ThonkerBot4.0SampleAfter/pyproject.toml
[tool.poetry]
name = "thonkerbot4-0sample"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "thonkerbot4"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {git = "https://github.com/Rapptz/discord.py.git"}
pyodbc = "^4.0.39"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
